Move msgpack serde tests into a sub-module and extract helpers
